% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AAprobC2}
\alias{AAprobC2}
\title{AA ambiguity probabilities.}
\usage{
AAprobC2(x, probs)
}
\arguments{
\item{x}{a heptovigdecimal integer (arity = 27).}

\item{probs}{a length-20 vector of logged probab.}
}
\description{
Find AA ambiguity probabilities.
}
