% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align2.R
\name{align}
\alias{align}
\title{Multiple sequence alignment.}
\usage{
align(sequences, type = "global", residues = "autodetect", gapchar = "-",
  DI = FALSE, ...)
}
\arguments{
\item{sequences}{a list of character vectors consisting of symbols from
the residue alphabet}

\item{residues}{one of the character strings "autodetect" (default), "aminos" or
"bases", or a case sensitive character vector matching the residue
alphabet. Specifying the symbol type ('aminos' or 'bases') can increase speed
for larger alignments. Note that setting \code{residues = "autodetect"} will not
detect rare residues that are not present in the alignment and thus will
not assign them emission probabilities.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.}
}
\description{
\code{align} finds the optimal alignment for a list of sequences using a hybrid
algorithm that involves a progressive alignment step followed by an iterative refinement
stage.
}

