% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alphadetect}
\alias{alphadetect}
\title{Detect residue alphabet.}
\usage{
alphadetect(sequences, residues = NULL, gapchar = "-", endchar = "?")
}
\arguments{
\item{sequences}{a character matrix or vector, or a list of character matrices
and/or vectors}
}
\description{
\code{"alphadetect"} performs checks on the format of the "residues" argument
to be passed to a variety of other functions. Single-element string arguments
such as "DNA" and "AA" are
converted to their respective alphabet in a character vector format.
}

