% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphid.R
\docType{package}
\name{aphid}
\alias{aphid}
\alias{aphid-package}
\title{aphid: an R package for analysis with profile hidden Markov models.}
\description{
Package 'aphid' contains functions for building profile hidden Markov models
for biological sequence analysis.
}
\section{aphid functions}{

}

