% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{arc}
\alias{arc}
\title{Geometric functions.}
\usage{
arc(x, y, radx, rady = radx, from = 0, to = 2 * pi, no.points = 100,
  fill = NULL, arrow = NULL, arrowsize = 0.08, code = 2, ...)
}
\description{
Geometric functions.
}

