% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{basesim}
\alias{basesim}
\title{Simulate the evolution of DNA}
\usage{
basesim(x, K, alphabet = c("a", "c", "g", "t"))
}
\arguments{
\item{x}{a vector of mode "character" containing the values "a", "c", "g", and "t". 
For \code{basesim}, the length of x must be one.}

\item{K}{the mutation rate in number of mutations per base pair per generation. 
Must be between 0 and 1.}

\item{n}{the number of generations to run the simulation for.}

\item{d}{an object of class \code{"dendrogram"}.}
}
\value{
an object of class \code{"dendrogram"} with a 'sequence' attribute at each node.
}
\description{
\code{basesim}, \code{seqsim}, and \code{treesim} simulate the evolution of DNA for a single base, 
a single sequence, and a set of sequences given a phylogenetic tree, respectively.
}
\examples{
set.seed(999)
myseq <- randomSequence(20)
mynewick <- "(A:1,(B:0.5,C:0.2):0.6);"
mydendrogram <- read.dendrogram(text = mynewick)
mysimulation <- treesim(myseq, K = 0.01, n = 100, d = mydendrogram)
mysimmatrix <- treesim.as.matrix(mysimulation)
myDNAbin <- ape::as.DNAbin(mysimmatrix)
mydistmat <- ape::dist.dna(myDNAbin, model = "JC69")
myhc <- hclust(mydistmat, "ave")
plot(as.dendrogram(myhc))

}

