% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaumWelch.R
\name{BaumWelch}
\alias{BaumWelch}
\alias{BaumWelch.HMM}
\alias{BaumWelch.PHMM}
\title{Iterative estimation of model parameters.}
\usage{
BaumWelch(x, y, maxiter = 100, deltaLL = 1e-07, logspace = FALSE,
  quiet = FALSE, pseudocounts = "Laplace")

\method{BaumWelch}{HMM}(x, y, maxiter = 100, deltaLL = 1e-07,
  logspace = FALSE, quiet = FALSE, modelend = FALSE,
  pseudocounts = "Laplace")

\method{BaumWelch}{PHMM}(x, y, maxiter = 100, deltaLL = 1e-07,
  logspace = FALSE, quiet = FALSE, pseudocounts = "Laplace")
}
\arguments{
\item{x}{an object of class \code{'HMM'} or \code{'PHMM'} specifying the
starting parameter values.}

\item{y}{a list of training sequences whose hidden states are unknown.}

\item{maxiter}{the maximum number of EM iterations before the process is terminated}

\item{deltaLL}{the change in log likelihood specified as the convergence threshold}

\item{logspace}{logical argument indicating whether the emission
and transmission probabilities povided for the model(s) are logged.}

\item{quiet}{logical argument indicating whether the iteration progress should
be suppressed (TRUE) or printed to the console (FALSE; default).}

\item{pseudocounts}{used to account for the possible absence of certain transition
and/or emission types in the training dataset.
either \code{'Laplace'} (adds one of each possible transition and emission type to the
training dataset; default), \code{'none'}, or a two-element list containing a matrix of
transition pseudocounts as its first element and a matrix of emission pseudocounts
as its second. If a list is supplied both matrices must have row and column names
according to the residues (column names of emission matrix) and states
(row and column names of the transition matrix and row names of the emission matrix).
The first row and column of the transition matrix must be 'BeginEnd'.}
}
\description{
\code{BaumWelch} uses a special case of the expectation-maximization (EM) algorithm
to find the locally-optimal parameters of a HMM or PHMM.
}
\seealso{
\code{\link{deriveHMM}} and \code{\link{derivePHMM}} for
maximum-likelihood parameter estimation when training sequence states are
known.
}

