% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustal.R
\name{clustalo}
\alias{clustalo}
\title{Multiple sequence alignment.}
\usage{
clustalo(x, quick = TRUE, pivot = FALSE, type = "semiglobal")
}
\arguments{
\item{x}{a named list of character vectors representing the sequences to be aligned.}

\item{quick}{logical value indicating whether the Wilbur Lipman k-tuples heuristic
should be used for the calculation of the pairwise distance matrix.}

\item{pivot}{logical value indicating whether pivot objects should be used
in mbed to allocate the most dissimilar seeds.}

\item{type}{a character string specifying whether the alignment should be
'global' (penalized end gaps) or 'semiglobal' (default; free end gaps).}
}
\value{
a character matrix of aligned sequences.
}
\description{
Align multiple DNA or protein
sequences using the clustal omega algorithm.
}

