% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivePHMM.R
\name{derive.PHMM}
\alias{derive.PHMM}
\alias{derive.PHMM.DNAbin}
\alias{derive.PHMM.AAbin}
\alias{derive.PHMM.list}
\alias{derive.PHMM.default}
\title{Derive a profile HMM from a multiple sequence alignment.}
\usage{
derive.PHMM(x, seqweights = "Gerstein", wfactor = 1, k = 5,
  residues = NULL, gapchar = "-", endchar = "?",
  pseudocounts = "background", logspace = TRUE, qa = NULL, qe = NULL,
  maxsize = NULL, inserts = "map", threshold = 0.5, lambda = 0,
  deltaLL = 1e-07, DI = FALSE, ID = FALSE, omit.endgaps = TRUE,
  name = NULL, description = NULL, compo = FALSE, consensus = FALSE,
  seeds = "random", refine = "Viterbi", maxiter = if (refine == "Viterbi")
  10 else 100, cpp = TRUE, quiet = FALSE, ...)

\method{derive.PHMM}{DNAbin}(x, seqweights = "Gerstein", wfactor = 1,
  k = 5, residues = NULL, gapchar = "-", endchar = "?",
  pseudocounts = "background", logspace = TRUE, qa = NULL, qe = NULL,
  maxsize = NULL, inserts = "map", threshold = 0.5, lambda = 0,
  deltaLL = 1e-07, DI = FALSE, ID = FALSE, omit.endgaps = TRUE,
  name = NULL, description = NULL, compo = FALSE, consensus = FALSE,
  seeds = "random", refine = "Viterbi", maxiter = if (refine == "Viterbi")
  10 else 100, cpp = cpp, quiet = FALSE, ...)

\method{derive.PHMM}{AAbin}(x, seqweights = "Gerstein", wfactor = 1,
  k = 5, residues = NULL, gapchar = "-", endchar = "?",
  pseudocounts = "background", logspace = TRUE, qa = NULL, qe = NULL,
  maxsize = NULL, inserts = "map", threshold = 0.5, lambda = 0,
  deltaLL = 1e-07, DI = FALSE, ID = FALSE, omit.endgaps = TRUE,
  name = NULL, description = NULL, compo = FALSE, consensus = FALSE,
  seeds = "random", refine = "Viterbi", maxiter = if (refine == "Viterbi")
  10 else 100, cpp = TRUE, quiet = FALSE, ...)

\method{derive.PHMM}{list}(x, seeds = "random", refine = "Viterbi",
  maxiter = if (refine == "Viterbi") 10 else 100, seqweights = "Gerstein",
  wfactor = 1, k = 5, residues = NULL, gapchar = "-",
  pseudocounts = "background", logspace = TRUE, qa = NULL, qe = NULL,
  maxsize = NULL, inserts = "map", lambda = 0, DI = FALSE, ID = FALSE,
  threshold = 0.5, deltaLL = 1e-07, omit.endgaps = TRUE, name = NULL,
  description = NULL, compo = FALSE, consensus = FALSE, cpp = TRUE,
  quiet = FALSE, ...)

\method{derive.PHMM}{default}(x, seqweights = "Gerstein", wfactor = 1,
  k = 5, residues = NULL, gapchar = "-", endchar = "?",
  pseudocounts = "background", logspace = TRUE, qa = NULL, qe = NULL,
  maxsize = NULL, inserts = "map", lambda = 0, threshold = 0.5,
  DI = FALSE, ID = FALSE, omit.endgaps = TRUE, name = NULL,
  description = NULL, compo = FALSE, consensus = FALSE, cpp = TRUE,
  quiet = FALSE)
}
\arguments{
\item{x}{a character matrix of aligned sequences.}

\item{seqweights}{either NULL (default; all sequences are given an equal
weight of 1), a numeric vector the same length as \code{x} representing
the sequence weights used to derive the model, or a character string giving
the method to derive the weights from the sequences. Currently only the
\code{"Gerstein"} method is supported (the default). For this method, a
tree is first created by k-mer counting (see \code{\link{kdistance}}),
and sequence weights are derived from the tree using the 'bottom up'
algorithm of Gerstein et al. (1994). The sum of these weights are equal
to the number of sequences in the alignment (so that mean(seqweights) = 1;
Note this does not need to be the case if providing weights as a numeric vector).}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the alignment), or a case sensitive character vector specifying the
residue alphabet (e.g. \code{c("A", "T", "G", "C")} for DNA).sequence weights are derived
Note that the former option can be slow for large character matrices;
therefore specifying the residue alphabet can increase speed in these cases.
Also note that the default setting \code{residues = NULL} will not
detect rare residues that are not present in the alignment, and thus will
not assign them emission probabilities.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.
Not required for \code{"DNAbin"} or \code{"AAbin"} objects.}

\item{pseudocounts}{the method used to account for the possible absence of certain
character states in the alignment. Currently only \code{"Laplace"}
and \code{"background"} or \code{"none"} are supported.
If \code{pseudocounts = "background"} and
some transition/emission types are absent from the alignment, Laplacean
pseudocounts are automatically added.}

\item{logspace}{logical indicating whether the emission and transition
probability on the returned model should be logged.}

\item{qa}{an optional named 9-element vector of background transition probabilities with
\code{dimnames(qa) = c("DD", "DM", "DI", "MD", "MM", "MI", "ID", "IM", "II")},
where M, I and D represent match, insert and delete states, respectively. If
\code{NULL}, background transition probabilities are generated from the alignment.}

\item{qe}{an optional named vector of background emission probabilities the
same length as the residue alphabet (i.e. 4 for nucleotides and 20 for
amino acids) and with corresponding names
(e.g. \code{c("A", "T", "G", "C")} for DNA).
If \code{NULL}, background emission probabilities are generated from the alignment.}

\item{inserts}{model construction method. Either 'threshold' (default),'map', 'inherited',
or if insert columns are to be specified manually, a logical vector of length ncol(x),
specifying the columns of x that should be designated
as inserts (\code{TRUE} for insert columns, \code{FALSE} for match states).
If \code{inserts = "threshold"} columns are assigned as insert states if the proportion of
gap characters in the column is higher than a specified threshold (defaults to 0.5).
If set to 'map' insert columns are found using the maximum \emph{a posteriori}
dynamic programming method outlined in Durbin et al. (1998). If set to 'inherited' the
function looks for an 'inserts' attribute from the alignment.}

\item{threshold}{the maximum proportion of gaps for an alignment column
to be considered a module in the PHMM (defaults to 0.5). Only applicable for
\code{inserts = "threshold"}. Note that the maximum \emph{a posteriori} method works
poorly for small alignments so the 'threshold' method is automatically used when
the number of sequences is fewer than 5.}

\item{lambda}{penalty parameter used to favour models with fewer match states. Equivalent
to the log of the prior probability of marking each column (Durbin et al. 1998, pg 124).
Only applicable for \code{inserts = "map"}.}
}
\value{
an object of class \code{"PHMM"}
}
\description{
\code{derive.PHMM} generates a profile HMM from a given multiple sequence alignment.
}
\details{
This function is the homologue of \code{hmmbuild} in HMMER3.1 (http://hmmer.org/) and
\code{modelfromalign} and \code{buildmodel} in the SAM package
(https://compbio.soe.ucsc.edu/sam.html).
}
\references{
Durbin R, Eddy SR, Krogh A, Mitchison G (1998) Biological
  sequence analysis: probabilistic models of proteins and nucleic acids.
  Cambridge University Press, Cambridge, United Kingdom.

  Gerstein M, Sonnhammer ELL, Chothia C (1994) Volume changes in protein evolution.
  \emph{Journal of Molecular Biology}, \strong{236}, 1067-1078.
}
