% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\title{Derive profile HMMs from sequence alignments.}
\usage{
derive(x, residues = "auto", gapchar = "-", method = "background",
  qe = NULL, qa = NULL)
}
\arguments{
\item{x}{a character matrix of aligned sequences.}

\item{residues}{one of the character strings 'auto' (default), aminos' or 
'bases', or a case sensitive character vector matching the residue 
alphabet. Specifying the symbol type ('aminos' or 'bases') can increase speed 
for larger alignments.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.}

\item{method}{the method used to account for the possible absence of certain
character states in the alignment columns. If \code{method = background} and 
some transition/emission types are absent from the alignment, Laplacean 
pseudocounts are automatically added.}

\item{qe}{an optional named vector of background emission probabilities the 
same length as the number of residues (i.e. 4 for nucleotides and 20 for 
amino acids) and with names corresponding to the residue alphabet 
(i.e. A C D E F G H I K L M N P Q R S T V W Y for amino acids or 
a c g t for nucleotides). If NULL, background emission probabilities are 
derived from the alignment.}

\item{qa}{an optional named 3 x 3 matrix of background transition probabilities with 
\code{dimnames(qa) = list(from = c('D', 'M', 'I'), to = c('D', 'M', 'I'))}, 
where M, I and D represent matches, inserts and deletions, respectively. If
NULL background transition probabilities are derived from the alignment.}
}
\description{
\code{derive} converts a multiple sequence alignment to a profile HMM.
}

