% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveHMM.R
\name{deriveHMM}
\alias{deriveHMM}
\title{Derive a HMM from training sequences.}
\usage{
deriveHMM(x, residues = "auto", states = "auto", modelend = FALSE,
  Apseudocounts = matrix(1, length(states), length(states), dimnames =
  list(from = states, to = states)), Epseudocounts = matrix(1,
  length(states[-1]), length(residues), dimnames = list(state = states[-1],
  residue = residues)))
}
\arguments{
\item{x}{a list of named character vectors representing residue emissions
from the model. The 'names' rattribute should represent the hidden state
that each residue was emitted from.}

\item{residues}{either \code{'auto'} (default), or a case
sensitive character vector matching the alphabet of residues
emitted by the model. Specifying the residue alphabet can increase speed
for larger training datasets. Note that setting \code{residues = 'auto'} will not
detect rare residues that are not present in the training data and thus will
not assign them emission probabilities. These should match or be a superset of
the unique residues present in the list of training sequences.}

\item{states}{either \code{'auto'} (default), or a case sensitive character
vector matching the hidden states. These should match or be a superset of
the unique states provided in the 'names' attributes in
the list of training
sequences.}

\item{modelend}{logical indicating whether transitions to the 'end' state should be
modeled. Defaults to FALSE.}

\item{method}{the method used to account for the possible absence of certain
character states in the alignment columns. Currently only \code{method = Laplace}
is supported.}
}
\description{
\code{deriveHMM} calculates the maximum likelihood hidden Markov model from
a list of training sequences, each a vector of emitted residues named according
the state from which they were emitted.
}

