% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivePHMM.R
\name{derivePHMM}
\alias{derivePHMM}
\title{Derive profile HMMs from sequence alignments.}
\usage{
derivePHMM(x, residues = "auto", gapchar = "-", method = "background",
  qe = NULL, qa = NULL, threshold = 0.5, DI = FALSE)
}
\arguments{
\item{x}{a character matrix of aligned sequences.}

\item{residues}{one of the character strings 'auto' (default), aminos' or
'bases', or a case sensitive character vector matching the residue
alphabet. Specifying the symbol type ('aminos' or 'bases') can increase speed
for larger alignments. Note that setting \code{residues = 'auto'} will not
detect rare residues that are not present in the alignment and thus will
not assign them emission probabilities.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.}

\item{method}{the method used to account for the possible absence of certain
character states in the alignment. Currently only \code{"Laplace"}
and \code{"background"} are supported. If \code{method = background} and
some transition/emission types are absent from the alignment, Laplacean
pseudocounts are automatically added.}

\item{qe}{an optional named vector of background emission probabilities the
same length as the residue alphabet (i.e. 4 for nucleotides and 20 for
amino acids) and with  corresponding names
(i.e. A C D E F G H I K L M N P Q R S T V W Y for amino acids or
a c g t for nucleotides). If NULL, background emission probabilities are
derived from the alignment.}

\item{qa}{an optional named 3 x 3 matrix of background transition probabilities with
\code{dimnames(qa) = list(from = c('D', 'M', 'I'), to = c('D', 'M', 'I'))},
where M, I and D represent matches, inserts and deletions, respectively. If
NULL background transition probabilities are derived from the alignment.}

\item{threshold}{the maximum proportion of gaps for an alignment column
to be considered a module in the PHMM (defaults to 0.5).
This is a simple heuristic; the option to assign columns to PHMM modules
by maximum \emph{a posteriori} (MAP) model construction will be added
in a later version.}
}
\description{
\code{derivePHMM} converts a multiple sequence alignment to a profile HMM.
}
\details{
This function is the homologue of \code{hmmbuild} in HMMER3.1 (http://hmmer.org/) and
\code{modelfromalign} and \code{buildmodel} in the SAM package
(https://compbio.soe.ucsc.edu/sam.html).
}

