% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DNAprobC2}
\alias{DNAprobC2}
\title{DNA ambiguity probabilities.}
\usage{
DNAprobC2(x, probs)
}
\arguments{
\item{x}{a pentadecimal integer (arity = 15).}

\item{probs}{a length-4 vector.}
}
\description{
Find DNA ambiguity probabilities.
}

