% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\alias{forward.HMM}
\alias{forward.PHMM}
\title{Full log likelihood.}
\usage{
forward(x, y, logspace = FALSE, global = TRUE)

\method{forward}{PHMM}(x, y, logspace = FALSE, global = TRUE)

\method{forward}{HMM}(x, y, logspace = FALSE)
}
\arguments{
\item{x}{an object of class \code{HMM} or \code{PHMM}, or a character vector.}
}
\description{
Forward and backward algorithms for calculating the full (log) probability
of a sequence given a HMM or PHMM.
}

