% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_HMM}
\alias{forward_HMM}
\title{Full probability of a sequence given a model.}
\usage{
forward_HMM(y, A, E, DNA = FALSE, AA = FALSE)
}
\arguments{
\item{y}{an integer vector with same arity as number of columns of E}
}
\description{
Implementation of the forward algorithm to caluclate the full (log) probability
of a sequence through a given a HMM or profile HMM.
}
