% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\alias{generate.HMM}
\alias{generate.PHMM}
\title{Generate random sequences from a model.}
\usage{
generate(x, size, logspace = "autodetect", gapchar = "-")

\method{generate}{HMM}(x, size, logspace = "autodetect")

\method{generate}{PHMM}(x, size, logspace = "autodetect", gapchar = "-")
}
\arguments{
\item{x}{an object of class \code{'HMM'} or \code{'PHMM'}.}

\item{size}{a non-negative integer representing the length of the output sequence
(if x is a HMM with zero probability of transitioning to the BeginEnd state),
or the maximum length of the output sequence (acts as a safeguard against overflow).}

\item{logspace}{logical argument indicating whether the emission and transition
probabilities of x are logged (base e; TRUE) or raw (FALSE). Alternatively, if
\code{logspace = "autodetect"} (default), the function will automatically detect
if the probabilities are in log space, returning an error if inconsistencies are found.
Note that choosing the latter option increases the computational
overhead; therefore specifying \code{TRUE} or \code{FALSE} can reduce the running time.}

\item{gapchar}{the character used to represent gaps in the output sequence
(\code{generate.PHMM} only).}
}
\value{
a named character vector giving the sequence of residues emitted by the model,
with "names" attribute representing the hidden states.
}
\description{
\code{generate} outputs an instance of a sequence of residues randomly emitted by a
given hidden Markov model.
}

