% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{kcount.DNA}
\alias{kcount.DNA}
\title{K-mer counting for DNA.}
\usage{
kcount.DNA(x, k = 5)
}
\arguments{
\item{x}{a list of DNA sequences in the "DNAbin" format.}

\item{k}{integer representing the length of the k-mers to be counted.}
}
\description{
\code{kcountDNA} takes a list of DNA sequences and returns a matrix
of k-mer counts with one row for each sequence and 4^k columns.
}
\details{
This function deals with ambiguities by assigning counts proportionally.
For example the motif ACRTG would assign the 5-mers ACATG and ACGTG counts of
0.5 each

This algorithm is of the order n * 4^k in memory and time so can be very
slow and memory hungry for large values of k (> 8).
}
\author{
Shaun P. Wilkinson
}
