% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{kdistance}
\alias{kdistance}
\title{K-mer distance matrix calculation.}
\usage{
kdistance(x, k = 5, alpha = "Dayhoff6", ...)
}
\arguments{
\item{x}{a list of sequences, can be an object of class \code{"DNAbin"} or \code{"AAbin"}.}

\item{k}{an integer specifying the size of the k-tuples.}

\item{alpha}{the residue alphabet.}

\item{...}{further arguments to be passed to \code{"dist"}.}
}
\description{
Computes the distance matrix ...
}

