% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logsum}
\alias{logsum}
\title{Sum of logged values}
\usage{
logsum(x)
}
\arguments{
\item{x}{a numeric vector of logged values (usually probabilities) to be summed over.}
}
\value{
a numerically stable calculation of \code{log(sum(exp(x)))}.
}
\description{
\code{logsum} takes a vector of logged values and returns the log of the sum of their
exponentials.
}

