% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivePHMM.R
\name{map}
\alias{map}
\title{Optimized profile HMM construction.}
\usage{
map(x, seqweights = NULL, residues = NULL, gapchar = "-", endchar = "?",
  pseudocounts = "background", lambda = 0, qa = NULL, qe = NULL,
  cpp = TRUE)
}
\arguments{
\item{x}{a matrix of aligned sequences.}

\item{seqweights}{either NULL (default; all sequences are given a weight of 1)
or a numeric vector of sequence weights used to
derive the model. The sum of these weights should be equal to the number of
sequences in the alignment, so that mean(seqweights) = 1.}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the alignment), or a case sensitive character vector specifying the
residue alphabet (e.g. A, C, G, T for DNA).
Note that the former option can be slow for large character matrices;
therefore specifying the residue alphabet can increase speed in these cases.
Also note that the default setting \code{residues = NULL} will not
detect rare residues that are not present in the alignment, and thus will
not assign them emission probabilities.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.
Not required for \code{"DNAbin"} or \code{"AAbin"} objects.}

\item{pseudocounts}{the method used to account for the possible absence of certain
character states in the alignment. Currently only \code{"Laplace"},
\code{"background"} and \code{"none"} are supported. If \code{pseudocounts = "background"} and
some transition/emission types are absent from the alignment, Laplacean
pseudocounts are automatically added.}

\item{lambda}{penalty parameter used to favour models with fewer match states. Equivalent
to the log of the prior probability of marking each column (Durbin et al. 1998, pg 124).}
}
\value{
a logical vector of length = ncol(x) indicating the columns to be assigned
as match states (\code{TRUE}) and those assigned as inserts (\code{FALSE}).

a logical vector indicating the columns of the input matrix that were allocated as
insert states
}
\description{
Assigns columns to insert states using the maximum \emph{a posteriori} method
outlined in Durbin et al. (1998).
}
\references{
Durbin...
}

