% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mbed}
\alias{mbed}
\title{Embed sequences as vectors of distances to \emph{n} seed sequences.}
\usage{
mbed(x, seeds = NULL, k = 5, alpha = "Dayhoff6")
}
\arguments{
\item{x}{a list of sequences, either character vectors or raw vectors in
"DNAbin" or "AAbin" format.}

\item{seeds}{integer vector indicating which sequences should be used as
the seed sequences.}

\item{k}{integer representing the length of the k-mers to be counted.}

\item{alpha}{the compression alphabet to use if the input sequence list
is an "AAbin" object. Currently the only supported options are "Dayhoff6"
and "none". Note that if \code{alpha = "none"} or the input sequences have
20 or more unique residues (e.g. protein sequences),
the value of k should be kept low (k < 4)
otherwise the number of k-mers to count will become excessively high.}
}
\description{
\code{embed.DNA} takes a list of sequences and returns a matrix of
  distances to a subset of seed sequences using the method outlined
  in Blacksheilds et al. (2010).
}
