% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior.HMM}
\alias{posterior.HMM}
\title{Posterior decoding.}
\usage{
\method{posterior}{HMM}(x, y, logspace = "autodetect")
}
\arguments{
\item{x}{an object of class \code{'HMM'} or \code{'PHMM'}.}

\item{y}{a character vector...}

\item{logspace}{logical argument indicating whether the emission and transition
probabilities of x are logged (base e; TRUE) or raw (FALSE). Alternatively, if
\code{logspace = "autodetect"} (default), the function will automatically detect
if the probabilities are in log space, returning an error if inconsistencies are found.
Note that choosing the latter option increases the computational
overhead; therefore specifying \code{TRUE} or \code{FALSE} can reduce the running time.}
}
\description{
Calculates the posterior probability of a sequence given a model.
}

