% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.PHMM}
\alias{print.HMM}
\alias{print.fullprob}
\alias{print.Viterbi}
\alias{print.mbed}
\title{Print summary methods.}
\usage{
\method{print}{PHMM}(x, ...)

\method{print}{HMM}(x, ...)

\method{print}{fullprob}(x, ...)

\method{print}{Viterbi}(x, ...)

\method{print}{mbed}(x, ...)
}
\arguments{
\item{x}{object of various classes.}

\item{...}{additional arguments.}
}
\value{
NULL (invisibly)
}
\description{
Print summary methods.
}
\author{
Shaun Wilkinson
}
