% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.PHMM}
\alias{read.PHMM}
\title{Import profile hidden Markov models into R.}
\usage{
read.PHMM(file = "", ...)
}
\arguments{
\item{file}{the name of the file from which to read the model.}

\item{...}{further arguments to be passed to \code{"scan"} .}
}
\value{
an object of class \code{"PHMM"}.
}
\description{
\code{read.PHMM} parses a HMMER3text file and creates an object of class
\code{"PHMM"}.
}
\references{
Finn, RD, Clements J & Eddy SR (2011) HMMER web server: interactive sequence similarity searching.
\emph{Nucleic Acids Research}. \strong{39}:W29–W37. \url{http://hmmer.org/}.
}
\seealso{
\code{\link{write.PHMM}} for writing PHMM objects in HMMER3 text format.
}

