% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readPHMM}
\alias{readPHMM}
\title{Import profile hidden Markov models into R.}
\usage{
readPHMM(file = "", ...)
}
\arguments{
\item{file}{the name of the file from which to read the model.}

\item{...}{further arguments to be passed to \code{"scan"}.}
}
\value{
an object of class \code{"PHMM"}.
}
\description{
The \code{readPHMM} function parses a HMMER3 text file into R and creates
  an object of class \code{"PHMM"}.
}
\details{
TBA
}
\examples{
  ## TBA
}
\references{
Finn, RD, Clements J & Eddy SR (2011) HMMER web server: interactive sequence
  similarity searching. \emph{Nucleic Acids Research}. \strong{39}:W29-W37.
  \url{http://hmmer.org/}.
}
\seealso{
\code{\link{writePHMM}} for writing PHMM objects in HMMER3 text format.
}
\author{
Shaun Wilkinson
}
