% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetPHMM.R
\name{subsetPHMM}
\alias{subsetPHMM}
\title{Subset an existing PHMM.}
\usage{
subsetPHMM(x, start, end)
}
\arguments{
\item{x}{an object of class \code{"PHMM"} to be subset.}

\item{start}{The first PHMM position to be inclused in the subset PHMM.}

\item{end}{The last PHMM position to be included in the subset PHMM.}
}
\value{
an object of class \code{"PHMM"}
}
\description{
The \code{subsetPHMM} function allows an existing PHMM to be subset
by profile position. This eliminates the need for the training of an additional, smaller
model if query sequences should be compared to only a subsection of an existing PHMM.
}
\examples{
## Small globin alignment data from Durbin et al (1998) Figure 5.3
data(globins)
## derive a profile hidden Markov model from the alignment
globins.PHMM <- derivePHMM(globins, residues = "AMINO", seqweights = NULL)
## subset positions 2-7 of the PHMM
short_globins.PHMM <- subsetPHMM(globins.PHMM, 2, 7)
}
\seealso{
\code{\link{derivePHMM}}
}
