% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{substitution_matrices}
\alias{substitution_matrices}
\title{Substutution matrices.}
\format{A list of...:
\describe{
  \item{PAM}{the PAM martices from... }
  \item{BLOSUM}{the BLOSUM matrices from...}
}}
\source{
\url{ftp://ftp.ncbi.nih.gov/blast/matrices/}
}
\usage{
substitution_matrices
}
\description{
A dataset containing several of the more popular substitution scoring matrices
for DNA and amino acids.
}
\keyword{datasets}
