% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdown.R
\name{topdown}
\alias{topdown}
\title{Build trees by recursive splitting}
\usage{
topdown(x, seeds = NULL, k = 5, weighted = TRUE)
}
\arguments{
\item{x}{a list of sequences, possibly an object of class
\code{"DNAbin"} or \code{"AAbin"}.}

\item{seeds}{an integer vector indicating which sequences should be used as
seeds for the clustering procedure (see details section).
If NULL the seeds are randomly selected from the input sequence list.}

\item{k}{integer representing the the k-mer size to be used for calculating
the distance matrix.}
}
\description{
\code{topdown} is used to cluster trees by recursive
partitioning, as opposed to agglomerative methods such as neighbour joining
and UPGMA. This is useful for large sequence sets (> 20, 000) since the
usual N x N distance matrix is not required. Instead, an N x log(N, 2)^2
distance matrix ('mbed') is derived from the input sequences, requiring substantially
less memory and time.
}
