% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transcount}
\alias{transcount}
\title{Count transition frequencies.}
\usage{
transcount(x, numbersystem)
}
\arguments{
\item{x}{an integer vector.}

\item{numbersystem}{an integer representing the numbering system of the input vector,
2 for binary, 3 for ternary, etc.}
}
\description{
Summation of transition frequencies in an integer vector.
}

