% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viterbi.R
\name{Viterbi}
\alias{Viterbi}
\alias{Viterbi.HMM}
\alias{Viterbi.PHMM}
\alias{Viterbi.default}
\title{Find the optimal sequence path recursively}
\usage{
Viterbi(x, y, qe = NULL, logspace = "autodetect", type = "semiglobal",
  offset = -0.1, d = 8, e = 2, S = NULL, itertab = NULL)

\method{Viterbi}{PHMM}(x, y, qe = NULL, logspace = "autodetect",
  type = "semiglobal", offset = -0.1, itertab = NULL)

\method{Viterbi}{HMM}(x, y, logspace = "autodetect")

\method{Viterbi}{default}(x, y, type = "semiglobal", d = 8, e = 2,
  S = NULL, itertab = NULL, offset = 0)
}
\arguments{
\item{x}{an object of class \code{HMM} or \code{PHMM}, or a character vector.}

\item{y}{a character vector consisting of residues emitted by the
HMM or PHMM.}

\item{qe}{an optional named vector of background residue frequencies. If NULL
background residue frequencies from the PHMMs are used. If these are not available
equal background residue frequencies are assumed.}

\item{logspace}{logical argument indicating whether the emission and transition
probabilities of x are logged (base e; TRUE) or raw (FALSE). Alternatively, if
\code{logspace = "autodetect"} (default), the function will automatically detect
if the probabilities are in log space, returning an error if inconsistencies are found.
Note that choosing the latter option increases the computational
overhead; therefore specifying \code{TRUE} or \code{FALSE} can reduce the running time.}

\item{type}{character string indicating whether insert and delete states
at the beginning and end of the path should count towards the final score
('global'; default), or not ('semiglobal'), or whether the highest scoring
sub-path should be returned ('local').}

\item{offset}{column score offset to specify level of greediness. Defaults to
-0.1 bits as recommended by Soding (2005).}

\item{d}{gap opening penalty for sequence vs. sequence alignment}

\item{e}{gap extension penalty for sequence vs. sequence alignment}

\item{S}{an optional scoring matrix with rows and columns named according
to the residue alphabet. Note that for local alignments scores for
mismatches should generally take negative values or else spurious
alignments could occur. If NULL matches are scored as 1 and
mismatches scored as -1.}

\item{itertab}{an optional two column matrix of row and column indices used
to iterate through a subset of the viterbi array.}
}
\description{
\code{Viterbi} finds the optimal path of a sequence through a HMM
or PHMM and returns its log-odds score.
}
\examples{
x <- c("H", "E", "A", "G", "A", "W", "G", "H", "E", "E")
y <- c("P", "A", "W", "H", "E", "A", "E")
Viterbi(x, y,  d = 8, e = 2)

}

