% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{Tree-based sequence weighting.}
\usage{
weight(x, method = "Gerstein")
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}.}

\item{method}{a character string indicating the weighting method to be used.
Currently only that of Gerstein et al. (1994) is supported
(\code{method = "Gerstein"}).}
}
\value{
a named vector of weights, the sum of which is equal to
the total number of sequences.
}
\description{
Takes a dendogram object and returns a vector of sequence weights following
the method of
Gerstein et al. (1994).
}
\examples{
den <- read.dendrogram(text = "(A:1,((B:0.5,C:0.2):0.2,(D:0.4,E:0.3):0.25):0.1);")
seqweights <- weight(den)

}

