% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window}
\alias{window}
\title{Define search space for dynammic programming.}
\usage{
window(x, y, arity = "autodetect", k = 4, w = 30, threshold = 5)
}
\arguments{
\item{x, y}{integer vectors representing sequences coded in a specific numbering scheme
(for example DNA residues A, T, G, C coded as 0, 1, 2, 3).}

\item{arity}{integer indicating the numbering system that x and y are coded in,
for example, 4 for DNA, and 20 for proteins.}

\item{threshold}{the number of standard deviations above the mean number of k-tuple matches
for all diagonals with k-tuple matches for a diagonal to be be considered significant.}
}
\value{
length 2 integer vector giving the bounds of the significant window space
    to fill the dynamic programming matrices.
}
\description{
Use the algorithm of Wilbur & Lipman (1983) to define windowspace for
downstream dynammic programming applications such as the Viterbi, forward and
backward algorithms.
}
\details{
TBA.
}
\references{
Wilbur WJ, Lipman DJ (1983) Rapid similarity searches of nucleic acid and
  protein data banks. \emph{Proc Natl Acad Sci USA}, \strong{10}, 197-206.
}
\author{
Shaun P. Wilkinson.
}
