% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.dendrogram}
\alias{write.dendrogram}
\title{Write dendrogram to Newick string}
\usage{
write.dendrogram(x, file = "", append = FALSE, strip.edges = FALSE,
  dec.places = 2)
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}.}

\item{file}{a character string naming a file or connection to write the output to.
If no file path is specified or \code{file = ""} the result will be printed to the console.}

\item{append}{a logical value indicating whether the output should be appended to the file.
If \code{append = FALSE} the contents of the file will be overwritten (the default setting).}

\item{strip.edges}{a logical value indicating whether edge weights should be 
removed from the output.}

\item{dec.places}{an integer incidating how many decimal places the edge weights should 
be rounded to.}
}
\description{
Writes a dendrogram object to a file or connection in Newick 
(New Hampshire) format.
}
\examples{
arrests.hc <- hclust(dist(USArrests[1:6,]), "ave")
arrests.den <- as.dendrogram(arrests.hc)
write.dendrogram(arrests.den)

}
\seealso{
\code{\link{read.dendrogram}} to create a \code{"dendrogram"} object from a 
text file.
}

