% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.PHMM}
\alias{write.PHMM}
\title{Export profile hidden Markov models as text.}
\usage{
write.PHMM(x, file = "", append = FALSE, form = "HMMER3", vers = "f")
}
\arguments{
\item{file}{the name of the file to write the model to.}

\item{append}{logical indicating whether the model text should be appended below any
existing text in the output file, or whether any existing text should be overwritten.
Defaults to FALSE.}

\item{form}{character string indicating the format in which to write the model.
Currently only HMMER3f is supported.}

\item{vers}{character string indicating the version of version of the format
in which to write the model. Currently only "f" is supported.}
}
\description{
\code{write.PHMM} takes an object of class
\code{"PHMM"} and writes it to a text file in HMMER3 format.
}
\references{
Finn, RD, Clements J & Eddy SR (2011) HMMER web server: interactive sequence similarity searching.
\emph{Nucleic Acids Research}. 39:W29–W37. \url{http://hmmer.org/}.
}
\seealso{
\code{\link{read.PHMM}} to parse a PHMM object from a HMMER3 text file.
}

