% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align2.R
\name{align}
\alias{align}
\title{Multiple sequence alignment.}
\usage{
align(sequences, type = "global", residues = "autodetect", gapchar = "-",
  DI = FALSE, ID = FALSE, refine = "Viterbi", quiet = FALSE, ...)
}
\arguments{
\item{sequences}{a list of character vectors consisting of symbols from
the residue alphabet}

\item{residues}{one of the character strings "autodetect" (default), "aminos" or
"bases", or a case sensitive character vector matching the residue
alphabet. Specifying the symbol type ('aminos' or 'bases') can increase speed
for larger alignments. Note that setting \code{residues = "autodetect"} will not
detect rare residues that are not present in the alignment and thus will
not assign them emission probabilities.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.}

\item{DI}{logical indicating whether delete -> insert transitions should be allowed
in the profile HMM. Defaults to FALSE, and not recommended for small training sets when
\code{refine = "BaumWelch"} due
to the tendency to converge to suboptimal local optima.}

\item{ID}{logical indicating whether insert -> delete transitions should be allowed
in the profile HMM. Defaults to FALSE, and similar to DI, not recommended for small
training sets when \code{refine = "BaumWelch"} due
to the tendency to converge to suboptimal local optima.}

\item{refine}{the method used to refine the model following progressive alignment.
Current supported methods are Viterbi training (\code{refine = "Viterbi"}) and
Baum Welch parameter optimization (\code{refine = "BaumWelch"}).}
}
\description{
\code{align} finds the optimal alignment for a list of sequences using a hybrid
algorithm that involves a progressive alignment step followed by an iterative refinement
stage.
}

