% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Pairwise alignment of sequences and/or sequence alignments.}
\usage{
align(x, y, d = 8, e = 2, S = NULL, qe = NULL, logspace = FALSE,
  type = "global", offset = -0.1, itertab = NULL, method = "background",
  residues = "auto", gapchar = "-")
}
\arguments{
\item{x, y}{character vectors or character matrices of aligned sequences.}

\item{d, e}{gap opening and gap extension penalties for pairwise sequence
alignment.}

\item{S}{an optional substitution matrix with \code{dimnames} attributes 
corresponding to the residue alphabet. If NULL matches are
scored as 1 and mismatches as -1.}

\item{type}{a character string specifying whether the alignment should be 
'global' (penalized end gaps), 'semiglobal' (default; free end gaps) or 
local (highest scoring subalignment).}
}
\value{
a character matrix of aligned sequences.
}
\description{
\code{align} uses the Viterbi algorithm to find the optimal alignment 
between two sequences, a sequence and an alignment, or two alignments.
}
\examples{
x <- c("H", "E", "A", "G", "A", "W", "G", "H", "E", "E")
y <- c("P", "A", "W", "H", "E", "A", "E")
z <- align(x, y)
align(x, z)

}

