% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Multiple sequence alignment.}
\usage{
align(sequences, type = "semiglobal", residues = NULL, gapchar = "-",
  DI = FALSE, ID = FALSE, refine = "Viterbi", quiet = TRUE,
  cpp = TRUE, ...)
}
\arguments{
\item{sequences}{a list of character vectors consisting of symbols from
the residue alphabet}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the list of sequences), or a case sensitive character vector specifying the
residue alphabet (e.g. c(A, C, G, T) for DNA).
Note that the former option can be slow for large lists of character vectors;
therefore specifying the residue alphabet can increase speed in these cases.
Also note that the default setting \code{residues = NULL} will not
detect rare residues that are not present in the sequence list, and thus will
not assign them emission probabilities.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.}

\item{DI}{logical indicating whether delete -> insert transitions should be allowed
in the profile HMM. Defaults to FALSE, and not recommended for small training sets when
\code{refine = "BaumWelch"} due
to the tendency to converge to suboptimal local optima.}

\item{ID}{logical indicating whether insert -> delete transitions should be allowed
in the profile HMM. Defaults to FALSE, and similar to DI, not recommended for small
training sets when \code{refine = "BaumWelch"} due
to the tendency to converge to suboptimal local optima.}

\item{refine}{the method used to refine the model following progressive alignment.
Current supported methods are Viterbi training (\code{refine = "Viterbi"}) and
Baum Welch parameter optimization (\code{refine = "BaumWelch"}).}

\item{...}{aditional arguments to pass to \code{"train"}.}
}
\description{
\code{align} finds the optimal alignment for a list of sequences using a hybrid
algorithm that involves a progressive alignment step followed by an iterative refinement
stage.
}

