% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align2phmm}
\alias{align2phmm}
\title{Align sequences to a profile HMM.}
\usage{
align2phmm(sequences, model, gapchar = "-", ...)
}
\arguments{
\item{...}{further arguments to be passed to \code{Viterbi}.}
}
\description{
Align sequences to a profile HMM.
}

