% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphid.R
\docType{package}
\name{aphid}
\alias{aphid}
\alias{aphid-package}
\title{aphid: an R package for analysis with profile hidden Markov models.}
\description{
Package 'aphid' contains functions for building profile hidden Markov models
for biological sequence analysis.
}
\section{aphid functions}{

  TBA
}

\references{
Durbin R, Eddy SR, Krogh A, Mitchison G (1998) Biological
  sequence analysis: probabilistic models of proteins and nucleic acids.
  Cambridge University Press, Cambridge, United Kingdom.

  Gerstein M, Sonnhammer ELL, Chothia C (1994) Volume changes in protein evolution.
  \emph{Journal of Molecular Biology}, \strong{236}, 1067-1078.
}
\author{
Shaun P. Wilkinson
}
