% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\alias{backward.HMM}
\alias{backward.PHMM}
\title{Backward algorithm.}
\usage{
backward(x, y, qe = NULL, logspace = "autodetect", odds = TRUE,
  type = "global", DI = TRUE, ID = TRUE)

\method{backward}{PHMM}(x, y, qe = NULL, logspace = "autodetect",
  odds = TRUE, type = "global", DI = TRUE, ID = TRUE)

\method{backward}{HMM}(x, y, logspace = "autodetect")
}
\arguments{
\item{x}{an object of class \code{PHMM} or \code{HMM}.}

\item{y}{a character vector representing a single instance of a sequence
hypothetically emitted by the model.}

\item{logspace}{logical argument indicating whether the emission and transition
probabilities of x are logged (base e; TRUE) or raw (FALSE). Alternatively, if
\code{logspace = "autodetect"} (default), the function will automatically detect
if the probabilities are in log space, returning an error if inconsistencies are found.
Note that choosing the latter option increases the computational
overhead; therefore specifying \code{TRUE} or \code{FALSE} can reduce the running time.}

\item{odds}{logical indicating whether the scores in th edynamic programming matrix
should be odds ratios (TRUE; default) or full (log) probabilities.
(\code{FALSE}; default) or if the log-odds score should be returned
(\code{TRUE}).}

\item{type}{character string indicating whether insert and delete states
at the beginning and end of the path should count towards the final score
('global'; default). Note that semiglobal and local models
are not currently supported in this version.}

\item{DI}{logical. should delete-insert transitions be allowed? Only applicable for
objects of class \code{"PHMM"}.}

\item{ID}{logical. should insert-delete transitions be allowed? Only applicable for
objects of class \code{"PHMM"}.}
}
\description{
Backward algorithm for calculating the full (log) probability
of a sequence given a hidden Markov model.
}

