% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\alias{backward.HMM}
\alias{backward.PHMM}
\title{Backward algorithm.}
\usage{
backward(x, obs, logspace = FALSE, odds = FALSE, type = "global")

\method{backward}{PHMM}(x, y, logspace = FALSE, odds = FALSE,
  type = "global")

\method{backward}{HMM}(x, y, logspace = FALSE)
}
\arguments{
\item{x}{an object of class \code{PHMM} or \code{HMM}.}

\item{logspace}{logical argument indicating whether the emission
and transmission probabilities povided for the model(s) are logged.}

\item{type}{character string indicating whether insert and delete states
at the beginning and end of the path should count towards the final score
('global'; default). Note that semiglobal and local models
are not currently supported in this version.}

\item{y}{a character vector representing a single instance of a sequence
hypothetically emitted by the model.}
}
\description{
Backward algorithm for calculating the full (log) probability
of a sequence given a HMM or PHMM.
}

