% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildPHMM}
\alias{buildPHMM}
\title{Build HMMs or profile HMMs}
\usage{
buildPHMM(E, A, qe, qa, n)
}
\arguments{
\item{E}{a matrix of emission probabilities}

\item{A}{a matrix of transition probabilities}

\item{qe}{a named vector of profile HMM background emission frequencies}

\item{qa}{a 3 x 3 matrix of profile HMM background transition frequencies}

\item{s}{a named vector of HMM start probabilities}
}
\description{
\code{buildPHMM} and \code{buildHMM} are functions for the manual creation of
objects of class \code{PHMM} and \code{HMM}, respectively.
}

