% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveHMM.R
\name{deriveHMM}
\alias{deriveHMM}
\title{Derive a HMM from training sequences.}
\usage{
deriveHMM(x, residues = "auto", states = "auto", modelend = FALSE,
  pseudocounts = "Laplace", logspace = FALSE)
}
\arguments{
\item{x}{a list of named character vectors representing residue emissions
from the model. The 'names' rattribute should represent the hidden state
that each residue was emitted from.}

\item{residues}{either \code{'auto'} (default), or a case
sensitive character vector matching the alphabet of residues
emitted by the model. Specifying the residue alphabet can increase speed
for larger training datasets. Note that setting \code{residues = 'auto'} will not
detect rare residues that are not present in the training data and thus will
not assign them emission probabilities. These should match or be a superset of
the unique residues present in the list of training sequences.}

\item{states}{either \code{'auto'} (default), or a case sensitive character
vector matching the hidden states. These should match or be a superset of
the unique states provided in the 'names' attributes in
the list of training
sequences.}

\item{modelend}{logical indicating whether transitions to the 'end' state should be
modeled. Defaults to FALSE.}

\item{pseudocounts}{used to account for the possible absence of certain transition
and/or emission types in the training dataset.
either \code{'Laplace'} (adds one of each possible transition and emission type to the
training dataset; default), \code{'none'}, or a two-element list containing a matrix of
transition pseudocounts as its first element and a matrix of emission pseudocounts
as its second. If a list is supplied both matrices must have row and column names
according to the residues (column names of emission matrix) and states
(row and column names of the transition matrix and row names of the emission matrix).
The first row and column of the transition matrix must be 'BeginEnd'.}
}
\description{
\code{deriveHMM} calculates the maximum likelihood hidden Markov model from
a list of training sequences, each a vector of residues named according
the state from which they were emitted.
}

