% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveHMM.R
\name{deriveHMM}
\alias{deriveHMM}
\title{Derive a HMM from training sequences.}
\usage{
deriveHMM(x, seqweights = NULL, residues = NULL, states = NULL,
  modelend = FALSE, pseudocounts = "background", logspace = FALSE,
  k = 1)
}
\arguments{
\item{x}{a list of named character vectors representing residue emissions
from the model. The 'names' attribute should represent the hidden state
that each residue was emitted from.}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the input sequences), or a case sensitive character vector specifying the
residue alphabet (e.g. A, C, G, T for DNA).
Note that the former option can be slow for large sequence lists;
therefore specifying the residue alphabet can increase speed in these cases.
Also note that the default setting \code{residues = NULL} will not
detect rare residues that are not present in the input sequences, and thus will
not assign them emission probabilities.}

\item{states}{either NULL (default; states are automatically detected from the 'names'
attributes of the input sequences), or a case sensitive character
vector matching the hidden states (these should match or be a superset of
the unique states provided in the 'names' attributes in the input sequences).}

\item{modelend}{logical indicating whether transitions to the 'end' state should be
modeled. Defaults to FALSE.}

\item{pseudocounts}{used to account for the possible absence of certain transition
and/or emission types in the input sequences.
either \code{'Laplace'} (adds one of each possible transition and emission type to the
training dataset; default), \code{'none'}, or a two-element list containing a matrix of
transition pseudocounts as its first element and a matrix of emission pseudocounts
as its second. If a list is supplied both matrices must have row and column names
according to the residues (column names of emission matrix) and states
(row and column names of the transition matrix and row names of the emission matrix).
The first row and column of the transition matrix must be 'BeginEnd'.}
}
\description{
\code{deriveHMM} calculates the maximum likelihood hidden Markov model from
a list of training sequences, each a vector of residues named according
the state from which they were emitted.
}

