% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\alias{forward.HMM}
\alias{forward.PHMM}
\title{Forward algorithm.}
\usage{
forward(x, y, qe = NULL, logspace = "autodetect", odds = TRUE,
  windowspace = "all", type = "global", eta = 0.01, DI = FALSE,
  ID = FALSE, cpp = TRUE)

\method{forward}{PHMM}(x, y, qe = NULL, logspace = "autodetect",
  type = "global", odds = TRUE, windowspace = "all", DI = FALSE,
  ID = FALSE, cpp = TRUE)

\method{forward}{HMM}(x, y, logspace = "autodetect", cpp = TRUE)
}
\arguments{
\item{x}{an object of class \code{PHMM} or \code{HMM}.}

\item{y}{a character vector representing a single instance of a sequence
hypothetically emitted by the model.}

\item{qe}{an optional named vector of background residue frequencies. If NULL
background residue frequencies from the PHMMs are used. If these are not available
equal background residue frequencies are assumed.}

\item{logspace}{logical argument indicating whether the emission and transition
probabilities of x are logged (base e; TRUE) or raw (FALSE). Alternatively, if
\code{logspace = "autodetect"} (default), the function will automatically detect
if the probabilities are in log space, returning an error if inconsistencies are found.
Note that choosing the latter option increases the computational
overhead; therefore specifying \code{TRUE} or \code{FALSE} can reduce the running time.}

\item{odds}{logical indicating whether the scores in the dynamic programming matrix
should be odds ratios (TRUE) or full logged probabilities (FALSE).}

\item{type}{character string indicating whether insert and delete states
at the beginning and end of the path should count towards the final score
('global'; default). Note that semiglobal and local models
are not currently supported in this version.}

\item{DI}{logical. should delete-insert transitions be allowed? Only applicable for
objects of class \code{"PHMM"}.}

\item{ID}{logical. should insert-delete transitions be allowed? Only applicable for
objects of class \code{"PHMM"}.}
}
\description{
Forward algorithm for calculating the full (log) probability or odds
of a sequence given a hidden Markov model or profile HMM.
}

