% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\alias{forward.PHMM}
\alias{forward.HMM}
\title{Forward algorithm.}
\usage{
forward(x, y, qe = NULL, logspace = "autodetect", odds = TRUE,
  windowspace = "all", type = "global", eta = 0.01, DI = FALSE,
  ID = FALSE, cpp = TRUE)

\method{forward}{PHMM}(x, y, qe = NULL, logspace = "autodetect",
  type = "global", odds = TRUE, windowspace = "all", DI = FALSE,
  ID = FALSE, cpp = TRUE)

\method{forward}{HMM}(x, y, logspace = "autodetect", cpp = TRUE)
}
\arguments{
\item{x}{an object of class \code{PHMM} or \code{HMM}.}

\item{y}{a vector of mode "character" or "raw" (a "DNAbin" or "AAbin"
object) representing a single sequence hypothetically emitted by
the model in \code{x}.}

\item{qe}{an optional named vector of background residue frequencies. If NULL
background residue frequencies from the PHMMs are used. If these are not available
equal background residue frequencies are assumed.}

\item{logspace}{logical argument indicating whether the emission and transition
probabilities of x are logged (base e; TRUE) or raw (FALSE). Alternatively, if
\code{logspace = "autodetect"} (default), the function will automatically detect
if the probabilities are in log space, returning an error if inconsistencies are found.
Note that choosing the latter option increases the computational
overhead; therefore specifying \code{TRUE} or \code{FALSE} can reduce the running time.}

\item{odds}{logical indicating whether the scores in the dynamic programming matrix
should be odds ratios (TRUE) or full logged probabilities (FALSE).}

\item{type}{character string indicating whether insert and delete states
at the beginning and end of the path should count towards the final score
('global'; default). Note that unlike \code{link{Viterbi}}, semiglobal
and local models are not currently supported.}

\item{DI}{logical. should delete-insert transitions be allowed? Only applicable for
objects of class \code{"PHMM"}.}

\item{ID}{logical. should insert-delete transitions be allowed? Only applicable for
objects of class \code{"PHMM"}.}
}
\value{
a list containing the score and dynamic programming arrays.
}
\description{
Calculate the full (log) probability or odds
of a sequence given a hidden Markov model or profile HMM with the
forward dynamic programming algorithm.
}
\details{
This function is a wrapper for a compiled C++ function that recursively
  fills a dynamic programming matrix with logged probabilities, and
  calculates the full (logged) probability of a sequence given a HMM or
  PHMM.
  For a thorough explanation of the backward, forward and Viterbi
  algorithms, see Durbin et al. (1998) chapters 3.2 (HMMs) and 5.4 (PHMMs).
}
\examples{
## Use of the forward algorithm for HMMs
## The dishonest casino example from Durbin et al. (1998) chapter 3.2
A <- matrix(c(0, 0, 0, 0.99, 0.95, 0.1, 0.01, 0.05, 0.9),
            nrow = 3) # transition probability matrix
dimnames(A) <- list(from = c("Begin", "Fair", "Loaded"),
                    to = c("Begin", "Fair", "Loaded"))
E <- matrix(c((1/6), (1/6), (1/6), (1/6), (1/6), (1/6),
              (1/10), (1/10), (1/10), (1/10), (1/10), (1/2)),
            nrow = 2, byrow = TRUE) # emission probability matrix
dimnames(E) <- list(states = c('Fair', 'Loaded'), residues = paste(1:6))
x <- structure(list(A = A, E = E), class = "HMM") # create hidden Markov model
data(casino)
forward(x, casino)

## Use of the forward algorithm for profile HMMs
## Small globin alignment data from Durbin et al. (1998) Figure 5.3
data(globins)
## derive a profile hidden Markov model from the alignment
globins.PHMM <- derivePHMM(globins, residues = "AMINO", seqweights = NULL)
## simulate a random sequence from the model
set.seed(999)
simulation <- generate(globins.PHMM, size = 20)
simulation ## "F" "S" "A" "N" "N" "D" "W" "E"
## calculate the full (log) probability of the sequence given the model
x <- forward(globins.PHMM, simulation, odds = FALSE)
x # -23.0586
## show dynammic programming array
x$array

}
\references{
Durbin R, Eddy SR, Krogh A, Mitchison G (1998) Biological
  sequence analysis: probabilistic models of proteins and nucleic acids.
  Cambridge University Press, Cambridge, United Kingdom.
}
\seealso{
\code{\link{backward}}, \code{\link{Viterbi}}.
}
\author{
Shaun P. Wilkinson
}
