% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\alias{forward.HMM}
\alias{forward.PHMM}
\alias{forward.default}
\title{Forward algorithm.}
\usage{
forward(x, y, logspace = "autodetect", odds = FALSE, type = "global")

\method{forward}{PHMM}(x, y, logspace = "autodetect", odds = FALSE,
  type = "global")

\method{forward}{HMM}(x, y, logspace = "autodetect")

\method{forward}{default}(x, y, type = "semiglobal", d = 8, e = 2,
  S = NULL, itertab = NULL, offset = 0)
}
\arguments{
\item{x}{an object of class \code{PHMM} or \code{HMM}.}

\item{y}{a character vector representing a single instance of a sequence
hypothetically emitted by the model.}

\item{logspace}{logical argument indicating whether the emission and transition
probabilities of x are logged (base e; TRUE) or raw (FALSE). Alternatively, if
\code{logspace = "autodetect"} (default), the function will automatically detect
if the probabilities are in log space, returning an error if inconsistencies are found.
Note that choosing the latter option increases the computational
overhead; therefore specifying \code{TRUE} or \code{FALSE} can reduce the running time.}

\item{type}{character string indicating whether insert and delete states
at the beginning and end of the path should count towards the final score
('global'; default). Note that semiglobal and local models
are not currently supported in this version.}
}
\description{
Forward algorithm for calculating the full log-odds
of a sequence given a HMM or PHMM and its random equivalent.
}

