% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{kdistance}
\alias{kdistance}
\title{K-mer distance matrix computation.}
\usage{
kdistance(x, k = 5, measure = "EDGAR04", residues = NULL, gap = "-",
  ...)
}
\arguments{
\item{x}{a matrix of aligned sequences or a list of unaligned sequences.
Accepted modes are "character" and "raw" (for "DNAbin" and "AAbin" objects).}

\item{k}{integer representing the k-mer size to be used for calculating
the distance matrix. Defaults to 5. Note that high values of k
may be slow to compute and use a lot of memory due to the large numbers
of calculations required, particularly when the residue alphabet is
also large.}

\item{measure}{a character string giving the k-mer distance measure
to be used. Currently the available options are \code{"EDGAR04"} (default;
see Edgar (2004) for details) and \code{"YANG08"} (the euclidean
distance measure described in Yang and Zhang (2008)).}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the sequences), a case sensitive character vector
specifying the residue alphabet, or one of the character strings
"RNA", "DNA", "AA", "AMINO". Note that the default option can be slow for
large lists of character vectors. Specifying the residue alphabet is therefore
recommended unless x is a "DNAbin" or "AAbin" object.}

\item{gap}{the character used to represent gaps in the alignment matrix
(if applicable). Ignored for \code{"DNAbin"} or \code{"AAbin"} objects.
Defaults to "-" otherwise.}

\item{...}{further arguments to be passed to \code{"as.dist"}.}
}
\value{
a distance matrix of class \code{"dist"}
}
\description{
Computes the matrix of k-mer distances between all pairwise comparisons
  of a set of sequences.
}
\details{
this function computes the N x N k-mer distance matrix,
  where N is the number of sequences in the dataset.
}
\examples{
  ## compute the k-mer distance for the woodmouse dataset in the ape package
  ## with a k-mer size of 4
  library(ape)
  data(woodmouse)
  woodmouse <- woodmouse[, 47:962] ## trim gappy ends for true global alignment
  woodmouse.dist <- kdistance(woodmouse, k = 4)
  ## compute and plot the UPGMA tree
  woodmouse.tree <- as.dendrogram(hclust(woodmouse.dist, "average"))
  plot(woodmouse.tree)
}
\references{
Edgar RC (2004) Local homology recognition and distance measures in
  linear time using compressed amino acid alphabets.
  \emph{Nucleic Acids Research}, \strong{32}, 380-385.

  Yang K, Zhang L (2008) Performance comparison between k-tuple distance
  and four model-based distances in phylogenetic tree reconstruction.
  \emph{Nucleic Acids Research}, \strong{36}, e33.
}
\seealso{
\code{\link{mbed}} for leaner distance matrices
}
\author{
Shaun Wilkinson
}
