% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logsum}
\alias{logsum}
\title{Sum of logged values.}
\usage{
logsum(x)
}
\arguments{
\item{x}{a vector of logged probabilities.}
}
\description{
\code{"logsum"} takes a vector of logged probabilities and returns.
}

