% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logsum}
\alias{logsum}
\title{Full log probability of sequence given model.}
\usage{
logsum(x)
}
\arguments{
\item{x}{an object of class \code{HMM} or \code{PHMM}.}
}
\description{
Implementation of the forward algorithm to caluclate the full (log) probability
of a sequence through a given a HMM or profile HMM.
}

