% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivePHMM.R
\name{map}
\alias{map}
\title{Optimal profile HMM construction.}
\usage{
map(x, residues = "autodetect", gapchar = "-",
  pseudocounts = "background", lambda = 0)
}
\arguments{
\item{x}{a character matrix of aligned sequences.}

\item{residues}{one of the character strings 'autodetect' (default), aminos' or
'bases', or a case sensitive character vector specifying the residue
alphabet. Specifying the symbol type ('aminos' or 'bases') can increase speed
for larger alignments. Note that setting \code{residues = 'autodetect'} will not
detect rare residues that are not present in the alignment and thus will
not assign them emission probabilities.}

\item{gapchar}{the character used to represent gaps in the alignment matrix.}

\item{pseudocounts}{the method used to account for the possible absence of certain
character states in the alignment. Currently only \code{"Laplace"},
\code{"background"} and \code{"none"} are supported. If \code{pseudocounts = "background"} and
some transition/emission types are absent from the alignment, Laplacean
pseudocounts are automatically added.}

\item{lambda}{penalty parameter used to favour models with fewer match states. Equivalent
to the log of the prior probability of marking each column (Durbin et al. 1998, pg 124).}
}
\value{
a logical vector of length = ncol(x) indicating the columns to be assigned
as match states (\code{TRUE}) and those assigned as inserts (\code{FALSE}).
}
\description{
Assigns columns to insert states using the maximum \emph{a posteriori} method
outlined in Durbin et al. (1998).
}
\references{
Durbin...
}

