% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mbed}
\alias{mbed}
\title{Convert sequences to vectors of distances to \emph{n} seed sequences.}
\usage{
mbed(x, seeds = NULL, k = 5, alpha = "Dayhoff6")
}
\arguments{
\item{x}{a list of sequences, either character vectors or raw vectors in
"DNAbin" or "AAbin" format.}

\item{seeds}{integer vector indicating which sequences should be used as
the seed sequences.}

\item{k}{integer representing the length of the k-mers to be counted.}

\item{alpha}{the compression alphabet to use if the input sequence list
is an "AAbin" object. Currently the only supported options are "Dayhoff6"
and "none". Note that if \code{alpha = "none"} or the input sequences have
20 or more unique residues (e.g. protein sequences),
the value of k should be kept low (k < 4)
otherwise the number of k-mers to count will become excessively high.}
}
\value{
returns a N x  matrix of class "embed". The returned object also has
  attributes including a matrix of k-mer counts with one row for each
  sequence and 4^k columns.
}
\description{
\code{embed.DNA} takes a list of sequences and returns a matrix of
  distances to a subset of seed sequences using the method outlined
  in Blacksheilds et al. (2010).
}
\details{
This function deals with ambiguities by assigning counts proportionally.
  For example the motif ACRTG would assign the 5-mers ACATG and ACGTG counts of
  0.5 each.
  This algorithm is order n * 4^k in memory and time complexity so can be very
  slow and memory hungry for large values of k (> 8).
}
\author{
Shaun P. Wilkinson
}
