% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.PHMM}
\alias{print.HMM}
\alias{print.DPA}
\title{Print summary methods.}
\usage{
\method{print}{PHMM}(x, ...)

\method{print}{HMM}(x, ...)

\method{print}{DPA}(x, ...)
}
\arguments{
\item{x}{object of various classes.}

\item{...}{additional arguments to be passed between methods.}
}
\value{
NULL (invisibly)
}
\description{
Print summary methods.
}
\author{
Shaun Wilkinson
}
