% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\alias{train.HMM}
\alias{train.PHMM}
\title{Viterbi training.}
\usage{
train(x, y, maxiter = 100, logspace = "autodetect", quiet = FALSE,
  modelend = FALSE, pseudocounts = "Laplace", fixqa = FALSE,
  fixqe = FALSE, inserts = "threshold", threshold = 0.5, lambda = 0,
  DI = TRUE, ID = TRUE)

\method{train}{PHMM}(x, y, maxiter = 100, logspace = "autodetect",
  quiet = FALSE, pseudocounts = "background", fixqa = FALSE,
  fixqe = FALSE, inserts = "map", threshold = 0.5, lambda = 0,
  DI = TRUE, ID = TRUE)

\method{train}{HMM}(x, y, maxiter = 100, logspace = "autodetect",
  quiet = FALSE, modelend = FALSE, pseudocounts = "Laplace")
}
\arguments{
\item{x}{the initial model with starting parameters to be updated.}

\item{y}{a list of training sequences (character vectors) whose hidden states are unknown.}
}
\description{
Update model parameters using a list of training sequences. Also known as the
segmental K-means algorithm (Juang & Rabiner 1990)
}
\references{
Juang B-H & Rabiner L R (1990) The segmental K-means
algorithm for estimating parameters of hidden Markov models.
IEEE transactions on Acoustics,
Speech...

Durbin...
}

