% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ViterbiC}
\alias{ViterbiC}
\alias{Viterbi_default}
\title{Optimal path of sequence through model.}
\usage{
Viterbi_default(x, y, type, d, e, S, windowspace, offset = 0)
}
\arguments{
\item{x}{an object of class \code{HMM} or \code{PHMM}, or a character vector.}

\item{y}{a character vector consisting of residues emitted by the
HMM or PHMM.}

\item{logspace}{logical argument indicating whether the emission
and transmission probabilities for the model(s) are logged.}
}
\description{
\code{ViterbiC} finds the optimal path of a sequence through a HMM
or PHMM and returns its log-odds score.
}
