% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\alias{weight.DNAbin}
\alias{weight.AAbin}
\alias{weight.list}
\alias{weight.dendrogram}
\alias{weight.default}
\title{Tree-based sequence weighting.}
\usage{
weight(x, method = "Gerstein", k = 5, residues = NULL, gapchar = "-")

\method{weight}{DNAbin}(x, method = "Gerstein", k = 5)

\method{weight}{AAbin}(x, method = "Gerstein", k = 5)

\method{weight}{list}(x, method = "Gerstein", k = 5, residues = NULL,
  gapchar = "-")

\method{weight}{dendrogram}(x, method = "Gerstein")

\method{weight}{default}(x, method = "Gerstein", k = 5, residues = NULL,
  gapchar = "-")
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}.}

\item{method}{a character string indicating the weighting method to be used.
Currently only that of Gerstein et al. (1994) is supported
(\code{method = "Gerstein"}).}
}
\value{
a named vector of weights, the sum of which is equal to
   the total number of sequences.
}
\description{
Takes a dendogram object and returns a vector of sequence weights following
the method of
Gerstein et al. (1994).
}
\details{
see Durbin et al. (1998) chapter 5.8
}
\examples{
data(woodmouse)
wood_dist <- kdistance(unalign(woodmouse))
wood_den <- as.dendrogram(hclust(wood_dist, method = "average"))
wood_weights <- weight(wood_den, method = "Gerstein")
}
