% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\alias{weight.DNAbin}
\alias{weight.AAbin}
\alias{weight.list}
\alias{weight.dendrogram}
\alias{weight.default}
\title{Tree-based sequence weighting.}
\usage{
weight(x, method = "Gerstein", k = 5, residues = NULL, gapchar = "-")

\method{weight}{DNAbin}(x, method = "Gerstein", k = 5)

\method{weight}{AAbin}(x, method = "Gerstein", k = 5)

\method{weight}{list}(x, method = "Gerstein", k = 5, residues = NULL,
  gapchar = "-")

\method{weight}{dendrogram}(x, method = "Gerstein")

\method{weight}{default}(x, method = "Gerstein", k = 5, residues = NULL,
  gapchar = "-")
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}, or a list or matrix of
sequences (possibly a "DNAbin" or "AAbin" object) from which to derive a
dendrogram via the \code{\link{topdown}} function.}

\item{method}{a character string indicating the weighting method to be used.
Currently only that of Gerstein et al. (1994) is supported
(\code{method = "Gerstein"}).}

\item{k}{integer representing the k-mer size to be used for calculating
the distance matrix used in tree-based sequence weighting. Defaults to
5. Note that high values of k (> 8) may be slow to compute and use a lot
of memory due to the large numbers of calculations required.}

\item{residues}{either NULL (default; emitted residues are automatically
detected from the sequences), a case sensitive character vector
specifying the residue alphabet, or one of the character strings
"RNA", "DNA", "AA", "AMINO". Note that the default option can be slow for
large lists of character vectors. Furthermore, the default setting
\code{residues = NULL} will not detect rare residues that are not present
in the sequences, and thus will not assign them emission probabilities
in the model. Specifying the residue alphabet is therefore
recommended unless x is a "DNAbin" or "AAbin" object.}

\item{gapchar}{the character used to represent gaps in the alignment matrix
(if applicable). Ignored for \code{"DNAbin"} or \code{"AAbin"} objects.
Defaults to "-" otherwise.}
}
\value{
a named vector of weights, the sum of which is equal to
   the total number of sequences (average weight = 1).
}
\description{
This function uses the agglomerative method of Gerstein et al. (1994) to
  weight sequences based on their relatedness, as derived from a
  phylogenetic tree.
}
\details{
For further details see Durbin et al. (1998) chapter 5.8.
}
\examples{
  ## weight the sequences in the woodmouse dataset from the ape package
  library(ape)
  data(woodmouse)
  woodmouse.weights <- weight(woodmouse)
  woodmouse.weights
}
\references{
Durbin R, Eddy SR, Krogh A, Mitchison G (1998) Biological
  sequence analysis: probabilistic models of proteins and nucleic acids.
  Cambridge University Press, Cambridge, United Kingdom.

  Gerstein M, Sonnhammer ELL, Chothia C (1994) Volume changes in protein evolution.
  \emph{Journal of Molecular Biology}, \strong{236}, 1067-1078.
}
\seealso{
\code{\link{topdown}}
}
\author{
Shaun Wilkinson
}
