% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichmax}
\alias{whichmax}
\title{Find the index of the maximum with ties broken at random.}
\usage{
whichmax(x, start = 1L)
}
\arguments{
\item{x}{a numeric or integer vector.}
}
\description{
Returns the location of the maximum value in a numeric or integer vector.
}
