% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.dendrogram}
\alias{write.dendrogram}
\title{Write dendrogram as a Newick string.}
\usage{
write.dendrogram(x, file = "", append = FALSE, edges = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"dendrogram"}.}

\item{file}{a character string naming a file or connection to write the output to.
If no file path is specified or \code{file = ""} the result will be printed to the console.}

\item{append}{a logical value indicating whether the output should be appended to the file.
If \code{append = FALSE} the contents of the file will be overwritten (the default setting).}

\item{edges}{a logical value indicating whether edge weights should be
included in the output string.}

\item{...}{further arguments to be passed to \code{format} to specify the numbering style
of the edge weights (assuming edges = TRUE).}
}
\description{
Writes a dendrogram object to a file or connection in Newick
(New Hampshire) format.
}
\examples{
arrests.hc <- hclust(dist(USArrests[1:6,]), "ave")
arrests.den <- as.dendrogram(arrests.hc)
write.dendrogram(arrests.den)

}
\seealso{
\code{\link{read.dendrogram}} to create a \code{"dendrogram"} object from a
text file.
}

